FDOS-177 code review updates
